refactor(glyph): extract the shared paragraph-to-engine encodings into one internal module
Constraint, region, style, and limit compilation plus minimal-text-edit
derivation now live in internal/engine-encoding.ts. The Three.js batch
binding keeps its behavior byte-for-byte while delegating to the same
encodings a framework-neutral host will use, so two hosts cannot drift
in what the engine is asked to flow.
