+++ packages/glyph/src/three/text.ts
@@ -48,6 +48,16 @@ import {
   type GlyphCaret,
   type GlyphPlacements,
 } from '../glyph-placement.js';
+import {
+  compileEngineGeometry,
+  engineLimits,
+  engineStyleValue,
+  minimalTextMutation,
+  normalizedColumns,
+  packedForeground,
+  replacedContent,
+  styledSpans,
+} from '../internal/engine-encoding.js';
 import { textFrameError, type TextFrameSubject } from './frame-error.js';
 import { ThreeTextRenderPlanExecutor } from './engine-plan-target.js';
 import {
@@ -703,7 +713,13 @@ class ThreeTextBatchBinding {
     }
     const properties = reconciler.properties(text);
     const content = properties.text as string;
-    const geometry = compileEngineGeometry(paragraph, properties.contentBox, 0, content.length);
+    const geometry = compileEngineGeometry(
+      paragraph.id,
+      paragraph.geometryRevision + 1,
+      properties.contentBox,
+      0,
+      content.length,
+    );
     const result = this.#session.measureParagraph(
       compileTextEngineFrameUpdate({
         sessionId: this.#session.handle,
@@ -901,7 +917,13 @@ class ThreeTextBatchBinding {
           }
         }
         if (semanticChanges & GEOMETRY_CHANGE) {
-          const geometry = compileEngineGeometry(paragraph, properties.contentBox, regions.length, content.length);
+          const geometry = compileEngineGeometry(
+            paragraph.id,
+            paragraph.geometryRevision + 1,
+            properties.contentBox,
+            regions.length,
+            content.length,
+          );
           constraints.push(geometry.constraint);
           regions.push(...geometry.regions);
         }
@@ -1250,27 +1272,6 @@ function compileEngineStyles<Technique extends AnyRasterTechnique>(
   return styles;
 }
 
-/**
- * The spans that state a style over text, in authored order.
- *
- * An empty span covers no cluster and so states nothing, but the engine's style resolution walks
- * the text offset by offset and cannot advance across a zero-width scope, so one reaching it fails
- * the whole frame (`style_state.rs`, `resolve`). Cluster resolution produces an empty span whenever
- * an edit hands a span's last cluster to the span before it, and keeps it in `Text.spans` so the
- * loss is visible and later indices do not shift; dropping it here is what keeps that record from
- * costing the paragraph. Style ids stay contiguous from the emitted order because the removal pass
- * that trims a shrunken style list counts on it.
- */
-function styledSpans<Technique extends AnyRasterTechnique>(
-  spans: readonly ParagraphSpan<Technique>[] | undefined,
-): readonly ParagraphSpan<Technique>[] {
-  // Only a collapsed span is dropped, and COMPILE TIME is where that belongs: the drop is a
-  // property of the engine style list, not of the caller's array, and doing it at `set()` would
-  // renumber every later span behind the caller's back. An INVERTED span never reaches here --
-  // `assertSpanRanges` throws it back at `set()`, where the stack still names the caller.
-  return spans === undefined ? [] : spans.filter((span) => span.start !== span.end);
-}
-
 function acquireEngineMaterial(
   coordinator: ThreeTextEngineCoordinator,
   material: ThreeTextMaterial | undefined,
@@ -1296,216 +1297,6 @@ function acquireEngineStack<Technique extends AnyRasterTechnique>(
   return lease.handle;
 }
 
-function engineStyleValue(
-  style: ParagraphStyle,
-  paint: GlyphPaintInput | undefined,
-  start: number,
-  end: number,
-  base: TextEngineStyleValue,
-): TextEngineStyleValue {
-  return {
-    ...base,
-    ...(style.fontSize === undefined ? {} : { fontSize: style.fontSize }),
-    ...(style.lineHeight === undefined ? {} : { lineHeight: style.lineHeight }),
-    ...(style.letterSpacing === undefined ? {} : { letterSpacing: style.letterSpacing }),
-    ...(style.wordSpacing === undefined ? {} : { wordSpacing: style.wordSpacing }),
-    ...(style.language === undefined ? {} : { language: style.language }),
-    ...(style.direction === undefined ? {} : { direction: style.direction }),
-    ...(style.features === undefined
-      ? {}
-      : {
-          features: style.features.map((feature) => ({
-            tag: feature.tag,
-            value: feature.value ?? 1,
-            start: feature.start ?? start,
-            end: feature.end ?? end,
-          })),
-        }),
-    ...(paint === undefined ? {} : { foregroundRgba: packedForeground(paint) }),
-    ...(style.decoration === undefined ? {} : { decoration: engineDecoration(style.decoration, paint) }),
-  };
-}
-
-function engineDecoration(
-  decoration: NonNullable<ParagraphStyle['decoration']>,
-  paint: GlyphPaintInput | undefined,
-): TextEngineDecoration {
-  if (decoration.style !== undefined && decoration.style !== 'solid') {
-    throw new TypeError(`'${decoration.style}' decoration lines are not implemented yet; only 'solid' is supported`);
-  }
-  return {
-    style: decoration.style ?? 'solid',
-    rgba: packedForeground(decoration.color === undefined ? (paint ?? {}) : { color: decoration.color }),
-    ...(decoration.underline === undefined ? {} : { underline: decoration.underline }),
-    ...(decoration.overline === undefined ? {} : { overline: decoration.overline }),
-    ...(decoration.lineThrough === undefined ? {} : { lineThrough: decoration.lineThrough }),
-    thickness: decoration.thickness ?? 0,
-    offset: decoration.offset ?? 0,
-  };
-}
-
-/**
- * Column region ids stride the high byte so every paragraph's extra columns
- * stay unique in the frame's region table while column zero keeps the
- * paragraph id itself — the single-column request bytes are unchanged.
- */
-const COLUMN_REGION_ID_STRIDE = 0x01_00_00_00;
-
-function normalizedColumns(contentBox: ParagraphContentBox | undefined): { count: number; gap: number } {
-  const columns = contentBox?.columns;
-  if (columns === undefined) return { count: 1, gap: 0 };
-  const gap = columns.gap ?? 0;
-  if (!Number.isSafeInteger(columns.count) || columns.count < 1 || columns.count > 16) {
-    throw new RangeError('contentBox columns count must be an integer between 1 and 16');
-  }
-  if (!Number.isFinite(gap) || gap < 0) {
-    throw new RangeError('contentBox columns gap must be a nonnegative finite number');
-  }
-  if (columns.count > 1 && contentBox?.width?.mode !== 'exact') {
-    throw new TypeError('contentBox columns require an exact width to derive the column measure');
-  }
-  // Ordered columns fill without balancing, so the column height is the only
-  // signal that advances flow into the next region: unbounded height would
-  // keep every line in the first column forever.
-  if (columns.count > 1 && contentBox?.height === undefined) {
-    throw new TypeError('contentBox columns require a bounded height to fill columns in order');
-  }
-  return { count: columns.count, gap };
-}
-
-function compileEngineGeometry(
-  paragraph: RetainedEngineParagraph,
-  contentBox: ParagraphContentBox | undefined,
-  regionStart: number,
-  textLength: number,
-): { readonly constraint: TextEngineConstraint; readonly regions: readonly TextEngineRegion[] } {
-  const width = axis(contentBox?.width);
-  const height = axis(contentBox?.height);
-  const columns = normalizedColumns(contentBox);
-  const inlineEnd = width.mode === 'unconstrained' ? 0x01_00_00_00 : width.size;
-  const blockEnd = height.mode === 'unconstrained' ? 0x01_00_00_00 : height.size;
-  const maxLines = contentBox?.maxLines ?? Math.max(1, textLength);
-  const geometryRevision = paragraph.geometryRevision + 1;
-  const columnWidth = (inlineEnd - columns.gap * (columns.count - 1)) / columns.count;
-  if (columns.count > 1 && columnWidth <= 0) {
-    throw new RangeError('contentBox columns and gap leave no positive column measure');
-  }
-  return {
-    constraint: {
-      paragraphId: paragraph.id,
-      flowThreadId: paragraph.id,
-      geometryRevision,
-      width: width.size,
-      height: height.size,
-      viewportBlockStart: 0,
-      viewportBlockEnd: blockEnd,
-      resumeBlockOffset: 0,
-      maxLines,
-      regionStart,
-      resumeCluster: 0,
-      regionCount: columns.count,
-      resumeRegion: 0,
-      widthMode: width.mode,
-      heightMode: height.mode,
-      wrap: contentBox?.wrap ?? 'word',
-      align: contentBox?.align ?? 'start',
-      overflow: contentBox?.overflow ?? 'visible',
-      blockAlign: 'start',
-      ...(contentBox?.firstLineIndent === undefined ? {} : { firstLineIndent: contentBox.firstLineIndent }),
-      ...(contentBox?.spaceBefore === undefined ? {} : { spaceBefore: contentBox.spaceBefore }),
-      ...(contentBox?.spaceAfter === undefined ? {} : { spaceAfter: contentBox.spaceAfter }),
-      ...(contentBox?.justify === undefined ? {} : { justify: contentBox.justify }),
-      ...(contentBox?.lastLine === undefined ? {} : { lastLine: contentBox.lastLine }),
-    },
-    regions: Array.from({ length: columns.count }, (_, column) => {
-      const inlineStart = column * (columnWidth + columns.gap);
-      const columnInlineEnd = column === columns.count - 1 ? inlineEnd : inlineStart + columnWidth;
-      return {
-        id: paragraph.id + COLUMN_REGION_ID_STRIDE * column,
-        geometryRevision,
-        transformIndex: paragraph.id,
-        shape: 'rectangle' as const,
-        exclusionStart: 0,
-        exclusionCount: 0,
-        writingMode: 'horizontal-tb' as const,
-        textOrientation: 'mixed' as const,
-        inlineStart,
-        blockStart: 0,
-        inlineEnd: columnInlineEnd,
-        blockEnd,
-        clipInlineStart: inlineStart,
-        clipBlockStart: 0,
-        clipInlineEnd: columnInlineEnd,
-        clipBlockEnd: blockEnd,
-      };
-    }),
-  };
-}
-
-function axis(value: ParagraphContentBox['width'] | undefined): {
-  readonly mode: 'unconstrained' | 'at-most' | 'exact';
-  readonly size: number;
-} {
-  if (value === undefined || value.mode === 'unconstrained') return { mode: 'unconstrained', size: 0 };
-  return { mode: value.mode, size: value.size };
-}
-
-function engineLimits(
-  paragraphCount: number,
-  textLength: number,
-  maximumParagraphTextLength: number,
-  regionCount: number,
-  maxOutputBytes: number,
-  mutationRecordCount = 0,
-): TextEngineFrameLimits {
-  return {
-    maxParagraphs: Math.max(1, paragraphCount),
-    maxClusters: Math.max(1, textLength * 2, mutationRecordCount),
-    // Rust applies this limit while composing each paragraph. Using aggregate batch text here makes every paragraph
-    // reserve enough line scratch for the whole TextGroup, multiplying retained memory by paragraph count.
-    maxLines: Math.max(1, maximumParagraphTextLength),
-    maxRegions: Math.max(1, regionCount),
-    maxExclusions: 1,
-    maxInlineObjects: 1,
-    maxSlotsPerBand: 8,
-    maxOutputBytes,
-  };
-}
-
-function packedForeground(paint: GlyphPaintInput): number {
-  const opacity = paint.opacity ?? 1;
-  if (!Number.isFinite(opacity) || opacity < 0 || opacity > 1) {
-    throw new RangeError('opacity must be in [0, 1]');
-  }
-  const input = paint.color ?? '#ffffff';
-  const rgba = typeof input === 'string' ? parseHexColorBytes(input) : linearColorBytes(input);
-  const alpha = Math.round(rgba[3] * opacity);
-  return (rgba[0] | (rgba[1] << 8) | (rgba[2] << 16) | (alpha << 24)) >>> 0;
-}
-
-function parseHexColorBytes(value: string): readonly [number, number, number, number] {
-  const match = /^#([0-9a-f]{6}|[0-9a-f]{8})$/iu.exec(value);
-  if (match === null) throw new TypeError('colors must be #rrggbb, #rrggbbaa, or linear RGBA');
-  const hex = match[1]!;
-  return [
-    Number.parseInt(hex.slice(0, 2), 16),
-    Number.parseInt(hex.slice(2, 4), 16),
-    Number.parseInt(hex.slice(4, 6), 16),
-    hex.length === 8 ? Number.parseInt(hex.slice(6), 16) : 255,
-  ];
-}
-
-function linearColorBytes(color: readonly number[]): readonly [number, number, number, number] {
-  if (color.length !== 4 || color.some((value) => !Number.isFinite(value) || value < 0 || value > 1)) {
-    throw new TypeError('linear RGBA colors must contain four finite channels in [0, 1]');
-  }
-  const srgbByte = (value: number): number => {
-    const srgb = value <= 0.003_130_8 ? value * 12.92 : 1.055 * value ** (1 / 2.4) - 0.055;
-    return Math.round(srgb * 255);
-  };
-  return [srgbByte(color[0]!), srgbByte(color[1]!), srgbByte(color[2]!), Math.round(color[3]! * 255)];
-}
-
 function releaseStackLeases(leases: readonly ThreeTextEngineStackLease[]): void {
   for (const lease of leases) lease.release();
 }
@@ -1517,49 +1308,6 @@ function releaseMaterialLeases(leases: readonly ThreeTextMaterialLease[]): void 
 function ownPublication(publication: TextEnginePublication): TextEnginePublication {
   const bytes = publication.bytes.slice();
   return { ...publication, bytes, memoryBuffer: bytes.buffer, memoryGrew: false };
-}
-
-/**
- * Replacement text carries its own formatting: a literal brings its spans and a
- * plain string brings none. Retaining the previous spans would reinterpret them
- * against unrelated text, so an update that replaces text without stating spans
- * clears the ones it replaced.
- */
-function replacedContent<Technique extends AnyRasterTechnique>(update: TextUpdate<Technique>): TextUpdate<Technique> {
-  if (!('text' in update) || 'spans' in update) return update;
-  return { ...update, spans: [] } as TextUpdate<Technique>;
-}
-
-function minimalTextMutation(previous: string, next: string): PendingTextMutation | undefined {
-  if (previous === next) return undefined;
-  const shared = Math.min(previous.length, next.length);
-  let start = 0;
-  while (start < shared) {
-    const previousCodePoint = previous.codePointAt(start)!;
-    if (previousCodePoint !== next.codePointAt(start)) break;
-    start += previousCodePoint > 0xffff ? 2 : 1;
-  }
-  let previousEnd = previous.length;
-  let nextEnd = next.length;
-  while (previousEnd > start && nextEnd > start) {
-    const previousStart = previousScalarStart(previous, previousEnd);
-    const nextStart = previousScalarStart(next, nextEnd);
-    if (previous.codePointAt(previousStart) !== next.codePointAt(nextStart)) break;
-    previousEnd = previousStart;
-    nextEnd = nextStart;
-  }
-  return {
-    start,
-    deleteCount: previousEnd - start,
-    insert: next.slice(start, nextEnd),
-  };
-}
-
-function previousScalarStart(value: string, end: number): number {
-  const last = end - 1;
-  const unit = value.charCodeAt(last);
-  const previous = value.charCodeAt(last - 1);
-  return unit >= 0xdc00 && unit <= 0xdfff && last > 0 && previous >= 0xd800 && previous <= 0xdbff ? last - 1 : last;
 }
 
 function classifySemanticChanges<Technique extends AnyRasterTechnique>(update: TextUpdate<Technique>): number {
